httputil: move signed client creation
